Add new folders and .sequelizerc
Added the main folder src, app and database. Moved all the existing folders, index.html and server.js to src; moved the models and controllers folder to app; and moved the migrations folder to database. Changed all the relevant path names to correspond to the new file structure and created a .sequelizerc file to change the path for sequelize.