Corrected paths and database config
I corrected some of the paths in package.json and routes.js because I forgot to add "app/". Also added some config data to directly refer to the database because it was throwing an error: The server does not support SSL connections. These were commented for now until further investigation.