Change timestamp columns to snake_case
Changed all the column names for createdAt and updatedAt to created_at and updated_at in the migrations, and also added a property that changes the camelCase attributes inside the models into snake_case when sequelize posts them to the database. It may cause a problem to the models because createdAt and updatedAt are properties that automatically receive a timestamp, so in case the models cause a problem, the new column names must be specified (check the manual: https://sequelize.org/master/manual/model-basics.html#timestamps )

Also, whoever wants to use these new migrations must first undo the previous migrations and migrate everything again. If there are seed, execute them after migrate.