Add Profissao controller and model and others
Added the model and controller for Profissao to see if it was possible to not have table columns for createdAt and updatedAt, and it worked! Also added a property to all models to change the tablename they reference to, created the /profissoes routes and changed the names for the test methods.