Remove tests that check number of rows and others
Removed all tests that would check if the final number of rows in each table corresponds to the number of correct queries. It would fail since all the tests were asynchronous and would sometimes interfere with each other's results.
Also removed the afterAll() that would delete all rows for the corresponding tables, for the same reasons above.